requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
click==7.1.2
    # via flask
flask==1.1.2
    # via -r requirements.in
itsdangerous==1.1.0
    # via flask
jinja2==2.11.2
    # via flask
markupsafe==1.1.1
    # via jinja2
python-dotenv==0.15.0
    # via -r requirements.in
werkzeug==1.0.1
    # via flask
